Abstract auth client into its own file.
