+++ pkg/vault/vault.go
@@ -11,26 +11,6 @@ import (
 	"go.opencensus.io/trace"
 )
 
-// AuthClient is a type that satifies the necesary authorization layer for a vault client.
-type AuthClient interface {
-	GetVaultToken(vc *vaultClient) (string, error)
-}
-
-func NewAuthClient(c *config) AuthClient {
-	switch {
-	case len(c.githubToken) > 0:
-		return NewGithubAuthClient()
-
-	case len(c.project) > 0:
-		return NewGcpAuthClient()
-	default:
-		log.Error("GetVaultToken: configuration error, one of [githubAuth, googleAuth] must be set to true")
-		os.Exit(1)
-	}
-
-	return nil
-}
-
 // NewVaultToken uses a github token or service account to get a vault auth token
 func NewVaultToken(vc *vaultClient) (string, error) {
 	if vc.config.traceEnabled {
